feat(routes): ajout des routes d’authentification et refactor de la route d’accueil (issue #13)
Fixes #13

- Création du module src/routes/authRoutes.js
  • POST /auth/register
  • POST /auth/login
  • DELETE /auth/delete/:id

- Extraction et refactor de la route d’accueil
  • Renommage accueil.js → accueilRoutes.js
  • Intégration propre dans app.js via app.use('/', accueilRoutes)

- Mise à jour de src/app.js pour monter les routes / et /auth

- Mise à jour de docs-dev/architecture.md
  • Ajout des modules accueilRoutes.js et authRoutes.js dans l’arborescence
  • Documentation de la séparation des responsabilités (issue 13)
  • Ajout de la section 2.1.3 (routes d’accueil et d’authentification)

Clôture de l’issue #13.
